Separate execution into their own types
